+++ source/win-main/file-manager/util/match-query.ts
@@ -26,11 +26,12 @@ type anyDescriptor = CodeFileMeta|DirMeta|MDFileMeta
  * descriptors (Codefiles, Directories, Markdown files) against the given query.
  *
  * @param   {string}    query                   The query string to match against.
+ * @param   {boolean}   includeTitle            Whether or not to include titles
  * @param   {boolean}   includeH1               Whether or not to include headings level 1
  *
  * @return  {(item: anyDescriptor) => boolean}  The filter function. Takes a descriptor as its only argument.
  */
-export default function matchQuery (query: string, includeH1: boolean): (item: anyDescriptor) => boolean {
+export default function matchQuery (query: string, includeTitle: boolean, includeH1: boolean): (item: anyDescriptor) => boolean {
   const queries = query.split(' ').map(q => q.trim()).filter(q => q !== '')
 
   // Returns a function that takes a Meta descriptor and returns whether it matches or not
@@ -63,7 +64,7 @@ export default function matchQuery (query: string, includeH1: boolean): (item: a
       const hasTitle = hasFrontmatter && 'title' in item.frontmatter
 
       // Does the frontmatter work?
-      if (hasTitle && String(item.frontmatter.title).toLowerCase().includes(q)) {
+      if (includeTitle && hasTitle && String(item.frontmatter.title).toLowerCase().includes(q)) {
         return true
       }